requirements.txt
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt[crypto]==5.5.1
